[Refactor] Divide Utils in different files
