+++ src/js/constants/validation.js
@@ -1,1 +1,2 @@
 export const MAX_TRIAL_NUMBER = 30;
+export const WAITING_TIME = 1000;
