+++ src/js/utils/carName.js
@@ -1,0 +1,6 @@
+export const splitingCarNames = (carNames) => {
+  const copyName = carNames;
+
+  if (!carNames.length) return [];
+  return copyName.split(',').filter((name) => name.trim());
+};
