+++ src/js/utils/random.js
@@ -1,0 +1,3 @@
+export const makeRandomNumber = (minValue = 0, maxValue = 9) => {
+  return Math.floor(Math.random() * maxValue + minValue);
+};
